backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Sumit Datta <sumitdatta@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
SQLAlchemy = "^1.4.26"
starlette = "^0.22.0"
uvicorn = "^0.15.0"
python-decouple = "^3.5"
databases = {extras = ["postgresql"], version = "^0.5.3"}
aiosqlite = "^0.17.0"
loguru = "^0.5.3"
aiohttp = "^3.7.4"
aioredis = "^2.0.0"
python-benedict = "^0.24.3"
sqlalchemy-migrate = "^0.13.0"
black = {extras = ["d"], version = "^22.10.0"}
python-dotenv = "^0.19.1"
PyJWT = "^2.3.0"
psycopg2 = "^2.9.3"
orjson = "^3.8.0"
google-auth = "^2.14.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
